Changed background image,  text-shadow color and increased area and added  box-shadow rule.
Changed background image to one that may be a bit easier to read text
on. Gave all the text in the form elements a lighter text-shadow color
to make it easier to read. Increased the box-shadow area for all the
fieldsets and gave the submit button a box-shadow rule to help tie it
in with the rest of the site.